Moved auth to a separate file, and refactored unit tests
