+++ tests/unit/test_db_models.py
@@ -210,7 +210,8 @@ class TestGet(TestUser):
         self.user = models.User(self.username, self.password, self.email)
         self.expected_result = {
             'username': self.username,
-            'email': self.email
+            'email': self.email,
+            'date_created': self.user.date_created.isoformat()
         }
 
     def test_get(self):
